refactor: `Board`의 정적 내부 클래스 `Position`을 외부로 분리
`Board` 클래스의 길이가 길어져 `Position`을 별도의 클래스로 분리하고 동일 패키지에 위치시킴.
